pyproject.toml
[tool.poetry]
name = "vanstore"
version = "0.1.0"
description = ""
authors = ["nock <ladrek13@mail.ru>"]
license = "MIT license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
